style: Refactor CSS imports and add new styles for components
