+++ src/components/SearchBar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./SearchBar.css";
+import "../styles/SearchBar.css";
 
 export default function SearchBar({ query, setQuery, onKeyDown }) {
   return (
