+++ src/styles/UpdateNotification.css
